third_party/WebKit/LayoutTests/http/tests/performance-timing/longtask/longtask-raf.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="resources/slowraf.js"></script>
<script>

async_test(function (t) {
    var observer = new PerformanceObserver(
        t.step_func(function (entryList) {
            var entries = entryList.getEntries();
            assert_equals(entries.length, 1,
                "Exactly one entry is expected.");
            var longtask = entries[0];
            assert_equals(longtask.entryType, "longtask");
            assert_equals(longtask.name, "self");
            assert_greater_than(longtask.duration, 50);

            // Assert the TaskAttributionTiming entry in attribution.
            assert_equals(longtask.attribution.length, 1,
                "Exactly one attribution entry is expected");
            var attribution = longtask.attribution[0];
            assert_equals(attribution.entryType, "taskattribution");
            assert_equals(attribution.name, "script");
            assert_equals(attribution.duration, 0);
            assert_equals(attribution.startTime, 0);

            observer.disconnect();
            t.done();
        })
    );
    observer.observe({entryTypes: ["longtask"]});

}, "Performance longtask entries are observable");

</script>
